fix: home changes according to the user
