+++ src/components/features/users/responsible/edit/ResponsibleEditForm.tsx
@@ -1,3 +1,4 @@
+import { FancyMultiSelect } from "@/components/forms/MultiSelect";
 import { Form } from "@/components/forms/Root";
 import { useUser } from "@/hooks/User/useUser";
 import api from "@/utils/api";
@@ -14,14 +15,15 @@ const formSchema = z.object({
     .min(2, { error: "Nome completo deve ter pelo menos 2 caracteres" }),
   email: z.email({ error: "Digite um email válido" }),
   escolaId: z.string().optional(),
+  usersIds: z.array(z.string()).optional(),
 });
 
-type ResponsableFormProps = {
+type ResponsibleFormProps = {
   id: number;
   onSuccess: () => void;
 };
 
-export function ResponsableEditForm({ id, onSuccess }: ResponsableFormProps) {
+export function ResponsibleEditForm({ id, onSuccess }: ResponsibleFormProps) {
   const form = useForm<z.infer<typeof formSchema>>({
     resolver: zodResolver(formSchema),
   });
@@ -29,6 +31,12 @@ export function ResponsableEditForm({ id, onSuccess }: ResponsableFormProps) {
   const [schools, setSchools] = useState<{ id: number; nome: string }[] | null>(
     null,
   );
+  const [students, setStudents] = useState<
+    { id: number; nome_completo: string; email: string }[] | null
+  >(null);
+  const [allStudents, setAllStudents] = useState<
+    { id: number; nome_completo: string; email: string }[] | null
+  >(null);
 
   useEffect(() => {
     const fetchUserData = async () => {
@@ -38,6 +46,7 @@ export function ResponsableEditForm({ id, onSuccess }: ResponsableFormProps) {
         if (response.status === 200) {
           const userData = {
             ...response.data,
+            escolaId: String(response.data.escolaId),
           };
           form.reset(userData);
         }
@@ -62,24 +71,65 @@ export function ResponsableEditForm({ id, onSuccess }: ResponsableFormProps) {
         }
       }
     };
+    const fetchRelations = async () => {
+      try {
+        const response = await api.get(`/responsible/list/${id}/students`);
+        if (response.status == 200) {
+          const users = response.data;
+          setStudents(users);
+          const originalIds = users.map((user: { id: number }) =>
+            String(user.id),
+          );
+          form.setValue("usersIds", originalIds);
+        }
+      } catch (error) {
+        console.log(error);
+      }
+    };
 
-    fetchUserData();
+    const fetchAllStudents = async () => {
+      try {
+        const response = await api.get(
+          `/user/list?type=Aluno${user?.perfil == "Admin" ? "" : `&escolaId=${user?.escola?.id}`}`,
+        );
+        if (response.status == 200) {
+          const users = response.data;
+          setAllStudents(users);
+        }
+      } catch (error) {
+        console.log(error);
+      }
+    };
+
+    fetchRelations();
+    fetchAllStudents();
     if (user?.perfil == "Admin") {
       fetchSchools();
     }
-  }, [id, form, user?.perfil]);
+    fetchUserData();
+  }, [id, form, user?.perfil, user?.escola?.id]);
 
   const onSubmit = async (data: z.infer<typeof formSchema>) => {
-    const payload = Object.fromEntries(
-      Object.entries(data).filter(
+    const userData = {
+      nome_completo: data.nome_completo,
+      email: data.email,
+      escolaId: data.escolaId,
+    };
+
+    const userPayload = Object.fromEntries(
+      Object.entries(userData).filter(
         ([_, value]) => value !== undefined && value !== "",
       ),
     );
 
     try {
-      const response = await api.put(`/user/update/${id}`, payload);
+      const responseUser = await api.put(`/user/update/${id}`, userPayload);
 
-      if (response.status === 200) {
+      await api.put(`/responsible/update/${id}`, {
+        usersIds: data.usersIds?.map((id) => Number(id)) || [],
+      });
+
+      if (responseUser.status === 200) {
         onSuccess();
       }
     } catch (error) {
@@ -123,7 +173,7 @@ export function ResponsableEditForm({ id, onSuccess }: ResponsableFormProps) {
             <Form.Input
               {...field}
               label="Nome Completo"
-              placeholder="Modificar nome completo"
+              placeholder="Mudar nome do(a) responsável"
             />
           )}
         />
@@ -144,12 +194,32 @@ export function ResponsableEditForm({ id, onSuccess }: ResponsableFormProps) {
             name="escolaId"
             render={({ field }) => (
               <Form.Select
-                {...field}
+                defaultValue={String(field.value)}
+                onChange={field.onChange}
                 label="Escola"
                 placeholder="Selecione a Escola"
                 options={schools.map((school) => ({
                   value: String(school.id),
                   label: school.nome,
+                }))}
+              />
+            )}
+          />
+        )}
+        {allStudents && students && (
+          <Form.Field
+            form={form}
+            name="usersIds"
+            render={({ field }) => (
+              <FancyMultiSelect
+                onSelect={field.onChange}
+                data={allStudents.map(({ id, email }) => ({
+                  value: String(id),
+                  label: email,
+                }))}
+                preSelectedData={students.map(({ id, email }) => ({
+                  value: String(id),
+                  label: email,
                 }))}
               />
             )}
