+++ src/pages/Dashboard/Dashboard.tsx
@@ -7,8 +7,8 @@ import { InfoBadge } from "../../components/utils/InfoBadge/InfoBadge";
 import "./Dashboard.css";
 
 import { RegisterSchoolUserModal } from "@/components/features/users/schoolUser/create/SchoolUserCreateModal";
-import { Footer } from "@/components/Footer/Footer";
-import { Header } from "@/components/Header/Header";
+import { Footer } from "@/components/footer/Footer";
+import { Header } from "@/components/header/Header";
 import { useUser } from "@/hooks/User/useUser";
 import { RegisterResponsableModal } from "../../components/features/users/responsible/create/ResponsibleCreateModal";
 import { RegisterSchoolModal } from "../../components/features/users/school/create/SchoolCreateModal";
